+++ javascript/modules/client/test/unit/context-plugins.test.ts
@@ -1,7 +1,7 @@
 // @ts-ignore
 declare const describe, it, beforeEach, expect, test;
 
-import { ContextErc20, ContextErc20Params } from "../../src";
+import { ContextPlugin, ContextPluginParams } from "../../src";
 import { Wallet } from "@ethersproject/wallet";
 import { JsonRpcProvider } from "@ethersproject/providers";
 import { activeContractsList } from "@aragon/core-contracts-ethers";
@@ -16,9 +16,9 @@ const web3endpoints = {
   failing: ["https://bad-url-gateway.io/"],
 };
 
-let contextParams: ContextErc20Params;
+let contextParams: ContextPluginParams;
 
-describe("ContextErc20 instances", () => {
+describe("ContextPlugin instances", () => {
   beforeEach(() => {
     contextParams = {
       network: "mainnet",
@@ -32,14 +32,14 @@ describe("ContextErc20 instances", () => {
     };
   });
   it("Should create an empty context", () => {
-    const context = new ContextErc20({});
+    const context = new ContextPlugin({});
 
-    expect(context).toBeInstanceOf(ContextErc20);
+    expect(context).toBeInstanceOf(ContextPlugin);
   });
   it("Should create an empty context and have default values", () => {
-    const context = new ContextErc20({});
+    const context = new ContextPlugin({});
 
-    expect(context).toBeInstanceOf(ContextErc20);
+    expect(context).toBeInstanceOf(ContextPlugin);
     expect(context.signer).toEqual(undefined);
     expect(context.dao).toEqual("");
     expect(context.daoFactoryAddress).toEqual(undefined);
@@ -47,9 +47,9 @@ describe("ContextErc20 instances", () => {
     expect(context.pluginAddress).toEqual("");
   });
   it("Should create a context and have the correct values", () => {
-    const context = new ContextErc20(contextParams);
+    const context = new ContextPlugin(contextParams);
 
-    expect(context).toBeInstanceOf(ContextErc20);
+    expect(context).toBeInstanceOf(ContextPlugin);
     expect(context.network).toEqual("mainnet");
     expect(context.signer).toBeInstanceOf(Wallet);
     expect(context.dao).toEqual("Dao");
@@ -64,22 +64,22 @@ describe("ContextErc20 instances", () => {
     contextParams.network = "notexistingnetwork";
 
     expect(() => {
-      new ContextErc20(contextParams);
+      new ContextPlugin(contextParams);
     }).toThrow();
   });
   it("Should create a context with invalid gas fee estimation factor and fail", () => {
     contextParams.gasFeeEstimationFactor = 1.1;
 
     expect(() => {
-      new ContextErc20(contextParams);
+      new ContextPlugin(contextParams);
     }).toThrow();
   });
   it("Should create a context with the correct DAOFactory address from the core-contracts-package", () => {
     contextParams.daoFactoryAddress = "";
     contextParams.network = "rinkeby";
-    const context = new ContextErc20(contextParams);
+    const context = new ContextPlugin(contextParams);
 
-    expect(context).toBeInstanceOf(ContextErc20);
+    expect(context).toBeInstanceOf(ContextPlugin);
     expect(context.network).toEqual("rinkeby");
     expect(context.daoFactoryAddress).toEqual(
       activeContractsList.rinkeby.DAOFactory,